Replace the english word "computed" with the frensh equivalent "calculée"
I this section you translated the first half of the concept "property" to "propriété" but not the second half "computed". I really don't know what is the idea behind this, but I think that "computed property" can be translated to "propriété calculée" and be more expressive for french speakers.